requirements.txt
aiohttp>=3.8.0
typing_extensions>=4.0.0
python-dotenv>=1.2.1
py-clob-client>=0.34.0
requests>=2.31.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0